Moved get_proposed_issuance_approval attribute to a mixin (it is used by three classes at this point)
